collectd: remove not needed fix
This was not a real fix but a workaround. It is no longer clear to me
why this was necessary. Deleting the patch restores the upstream
behaviour of the collected for the smart plugin. I have tested it and on
my system the hard disk to be monitored is recognised.

root@system ~ # cat /sys/class/block/sda/uevent
MAJOR=8
MINOR=0
DEVNAME=sda
DEVTYPE=disk

Signed-off-by: Florian Eckert <fe@dev.tdt.de>
(cherry picked from commit c62199ad69cf635b7ef48179cb0cc714697496c7)