gcc: update to allow compiling different versions
This is based on the toolchain GCC, and aims to share as much of its
Makefile and patches with that definition. The package requires one
additional patch:

(1) 003-dont-choke-when-building-32bit-on-64bit.patch, which fixes the
`error: size of array 'test_real_width' is negative` error that occurs
when building a 32-bit GCC on a 64-bit host. (Search the Internet for
examples of this error appearing.)

Signed-off-by: W. Michael Petullo <mike@flyn.org>
(cherry picked from commit 2a0ba0b4e388710246ac9398c4b15be831a97f78)
Signed-off-by: Josef Schlehofer <pepe.schlehofer@gmail.com>
[added gcc version 9.3.0, downgraded gcc to 10.2.0, removed gcc version
11.3.0 as it is now in sync with OpenWrt 21.02]