protobuf: fix 46e9f92
The cherry-pick done in 46e9f92 includes changing the build setup from
cmake to ninja, but it was overlooked that this was actually reverted in
03887a9.

The ninja build results in headers not being installed for the host pkg,
so protobuf-c/host can't be build.

This commit reverts the package back to cmake.

Signed-off-by: Sebastian Kemper <sebastian_ml@gmx.net>