attendedsysupgrade-common: update ASU CA pubkey
The old key was overwritten during a machine migration and the current
worker key is no longer valid. To create new valid worker certificates
this commit adds a new CA pubkey.

To update running OpenWrt devices via a client one needs to force
install it or upgrade to the latest version of
`attendedsysupgrade-common` - I'm sorry for the inconvenience.

This change is only related to the `usign` pubkey and `ucert`
certificate which is used to sign all created images on the server.

While at it, use INSTALL_DATA instead of INSTALL_BIN.

Signed-off-by: Paul Spooren <mail@aparcar.org>
(cherry picked from commit 050c749f4a2945219644081094e1647d82f98046)